fix(ci): restore workspace portability checks
